refactor: split files in project per feature
