+++ Kysect.GithubActivityAnalyzer.Sample/Program.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text.Json;
-using Kysect.GithubActivityAnalyzer.Services;
 
 namespace Kysect.GithubActivityAnalyzer.Sample
 {
