+++ Kysect.GithubActivityAnalyzer.WebDemo/Client/Pages/GetStudyGroup.razor
@@ -1,6 +1,6 @@
 ﻿@page "/getstudyGroup"
 @using Kysect.GithubActivityAnalyzer.WebDemo.Shared
-@using Kysect.GithubActivityAnalyzer.Models.Aggregations
+@using Kysect.GithubActivityAnalyzer.Aggregators.Models
 @inject HttpClient Http
 <h1>TestFunc</h1>
 <p>
