+++ Kysect.GithubActivityAnalyzer.WebDemo/Client/Pages/TestPage.razor
@@ -1,6 +1,6 @@
 ﻿@page "/testpage"
-@using Kysect.GithubActivityAnalyzer.Models.ApiResponses
 @using System.Text.Json
+@using Kysect.GithubActivityAnalyzer.ApiAccessor.ApiResponses
 @inject HttpClient Http
 <h1>TestFunc</h1>
 <p>
