+++ Kysect.GithubActivityAnalyzer/Aggregators/Models/MonthlyStatistics.cs
@@ -1,9 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using Kysect.GithubActivityAnalyzer.Services;
 
-namespace Kysect.GithubActivityAnalyzer.Models.Aggregations
+namespace Kysect.GithubActivityAnalyzer.Aggregators.Models
 {
     public class MonthlyStatistics
     {
